move to partition.rs
